Make Todo carry owned strings
This change accomodates the top level function to parse a file. We open a file
in a loop and read its contents to parse. But when the file goes out of scope,
the Todos are pointing to out of scope &str's and the compiler complains. With
this change, Todos are pretty much independent and can be passed around even
after closing the file they refer to.